app/requirements.txt
fastAPi
pandas
psycopg2-binary
unidecode
python-multipart
uvicorn
sqlalchemy
pytest